big push / likes

+++ Api/controllers/user.js
@@ -34,7 +34,7 @@ exports.login = (req, res, next) => {
       'SELECT * FROM users WHERE email= ?',
       userReq,
       function (_error, results, _fields) {
-        if (results.length > 0) {
+        if (results && results.length > 0) {
           bcrypt.compare(passReq, results[0].password).then((valid) => {
             if (!valid) {
               res
@@ -87,31 +87,73 @@ exports.getAllUsers = (req, res, next) => {
   )
 }
 
-exports.current = (req, res, next) => {
+exports.getUser = (req, res, next) => {
+  let userId;
+  if (req.params.id){
+    userId=req.params.id
+  }
+  else{
+    const token = req.headers.authorization.split(' ')[1]
+    const decodedToken = jwt.verify(token, config.token)
+    userId = decodedToken.userId
+  }  
   conn.query(
-    'SELECT idUSERS, username, isAdmin, email FROM users WHERE idUsers= ?',
-    req.userId,
+    "SELECT u.idUSERS, u.username, u.isAdmin, u.email, COUNT(DISTINCT m.idMESSAGES) AS 'nombre_message', COUNT(DISTINCT l.idMESSAGES) AS 'nombre_likes' FROM users u LEFT JOIN messages m ON u.idUSERS = m.idUSERS AND m.message_parent IS NULL LEFT JOIN likes l ON u.idUSERS = l.idUSERS WHERE u.idUSERS= ?",
+    userId,
     function (error, results, fields) {
       if (error) {
         return res.status(400).json(error)
       }
-      console.log(results)
+      if (results[0].idUSERS === null ){
+        return res.status(404).json("Utilisateur introuvable")
+      }
       return res.status(200).json({ user:results })
     }
   )
 }
 
-exports.deleteUser = (req, res, next) => {
+exports.getUserMessages = (req, res, next) => {
   conn.query(
-    `DELETE FROM users WHERE idUSERS=${req.params.id}`,
+    'SELECT * FROM messages WHERE idUSERS= ? AND message_parent IS NULL',
     req.params.id,
     function (error, results, fields) {
       if (error) {
         return res.status(400).json(error)
       }
-      return res
-        .status(200)
-        .json({ message: 'Votre compte a bien été supprimé !' })
+      return res.status(200).json({ msg:results })
     }
   )
+}
+
+exports.getUserLikes = (req, res, next) => {
+  conn.query(
+    'SELECT l.*,m.title FROM likes l LEFT JOIN messages m ON l.idMESSAGES=m.idMESSAGES WHERE l.idUSERS= ?',
+    req.params.id,
+    function (error, results, fields) {
+      if (error) {
+        return res.status(400).json(error)
+      }
+      return res.status(200).json({ likes:results })
+    }
+  )
+}
+
+exports.deleteUser = (req, res, next) => {
+  const token = req.headers.authorization.split(' ')[1]
+  const decodedToken = jwt.verify(token, config.token)
+  if((decodedToken.role === 'admin' && decodedToken.userId !== req.params.id ) || decodedToken.userId === req.params.id ){
+    conn.query(
+      `DELETE FROM users WHERE idUSERS=${req.params.id}`,
+      req.params.id,
+      function (error, results, fields) {
+        if (error) {
+          return res.status(400).json(error)
+        }
+        return res
+          .status(200)
+          .json({ message: 'Le compte a bien été supprimé !' })
+      }
+    )
+  }
+  return res.status(401).json({ message: 'Supression non autorisée' })
 }
